Create different submission files based on assignment
Canvas, SFU's main online course delivery software, does not use different folders for each assignment upon submission.
Because of this, all assignments end up being uploaded as "a-7.zip" or whatever it may be.

This will cause any first uploads of the assignment (not resubmissions. But those should be commented as such) to not include any appended numbers.